refactor(components): Phase 3 - content components under components/content/
- Move home, about, projects, blog to components/content/
- Update all imports to @/components/content/{home,about,projects,blog}
- Remove unused skeletons (featured-projects, blog-list, project-list)
- Cleaner separation: content/ for page content, ide/ for IDE chrome
