+++ apps/web/src/components/content/about/about-bio.tsx
