+++ apps/web/src/components/content/home/hero-bio.tsx
