+++ apps/web/src/components/content/home/hero-quote.tsx
