+++ apps/web/src/components/content/projects/project-card-details.tsx
